refactor: replace JPG background images with WEBP format for improved performance
