+++ devbyte-frontend/src/pages/auth/OtpVerification.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { ArrowLeft, CheckCircle } from "lucide-react";
 import { motion, AnimatePresence } from "framer-motion";
-import bgImage from "@/assets/images/auth-bg2.jpg";
+import bgImage from "@/assets/images/auth-bg2.WEBP";
 
 const OtpVerification = () => {
   const [otp, setOtp] = useState(["", "", "", "", "", ""]);
